Pull SpriteClient into seagulls engine
